Move cloud storage one level up
